+++ examples/master_querant.rs
@@ -165,6 +165,8 @@ fn main() -> GDResult<()> {
         "ut" => println!("{:#?}", ut::query(ip, port)),
         "bf1942" => println!("{:#?}", bf1942::query(ip, port)),
         "ss" => println!("{:#?}", ss::query(ip, port)),
+        "_gamespy3" => println!("{:#?}", gamespy::three::query(ip, port.unwrap(), None)),
+        "_gamespy3_vars" => println!("{:#?}", gamespy::three::query_vars(ip, port.unwrap(), None)),
         _ => panic!("Undefined game: {}", args[1]),
     };