+++ src/protocols/gamespy/mod.rs
@@ -1,7 +1,5 @@
-/// The implementation.
-pub mod protocol;
-/// All types used by the implementation.
-pub mod types;
+mod common;
+/// The implementations.
+pub mod protocols;
 
-pub use protocol::*;
-pub use types::*;
+pub use protocols::*;