+++ src/protocols/gamespy/protocol/mod.rs
@@ -1,2 +1,0 @@
-/// GameSpy 1
-pub mod one;