+++ src/protocols/gamespy/protocols/one/protocol.rs
@@ -1,7 +1,7 @@
 use crate::{
     bufferer::{Bufferer, Endianess},
     protocols::{
-        gamespy::{Player, Response},
+        gamespy::one::{Player, Response},
         types::TimeoutSettings,
     },
     socket::{Socket, UdpSocket},
@@ -9,6 +9,7 @@ use crate::{
     GDResult,
 };
 
+use crate::protocols::gamespy::common::has_password;
 use std::collections::HashMap;
 
 fn get_server_values(
@@ -173,21 +174,6 @@ fn extract_players(server_vars: &mut HashMap<String, String>, players_maximum: u
     Ok(players)
 }
 
-fn has_password(server_vars: &mut HashMap<String, String>) -> GDResult<bool> {
-    let password_value = server_vars
-        .remove("password")
-        .ok_or(GDError::PacketBad)?
-        .to_lowercase();
-
-    if let Ok(has) = password_value.parse::<bool>() {
-        return Ok(has);
-    }
-
-    let as_numeral: u8 = password_value.parse().map_err(|_| GDError::TypeParse)?;
-
-    Ok(as_numeral != 0)
-}
-
 /// If there are parsing problems using the `query` function, you can directly
 /// get the server's values using this function.
 pub fn query_vars(
@@ -209,6 +195,10 @@ pub fn query(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>
         .ok_or(GDError::PacketBad)?
         .parse()
         .map_err(|_| GDError::TypeParse)?;
+    let players_minimum = match server_vars.remove("minplayers") {
+        None => None,
+        Some(v) => Some(v.parse::<u8>().map_err(|_| GDError::TypeParse)?),
+    };
 
     let players = extract_players(&mut server_vars, players_maximum)?;
 
@@ -225,15 +215,11 @@ pub fn query(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>
         game_version: server_vars.remove("gamever").ok_or(GDError::PacketBad)?,
         players_maximum,
         players_online: players.len(),
-        players_minimum: server_vars
-            .remove("minplayers")
-            .unwrap_or_else(|| "0".to_string())
-            .parse()
-            .map_err(|_| GDError::TypeParse)?,
+        players_minimum,
         players,
         tournament: server_vars
             .remove("tournament")
-            .unwrap_or_else(|| "true".to_string())
+            .unwrap_or("true".to_string())
             .to_lowercase()
             .parse()
             .map_err(|_| GDError::TypeParse)?,