+++ src/protocols/gamespy/protocols/three/types.rs
@@ -8,16 +8,19 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct Player {
     pub name: String,
-    pub team: u8,
-    /// The ping from the server's perspective.
+    pub score: i32,
     pub ping: u16,
-    pub face: String,
-    pub skin: String,
-    pub mesh: String,
-    pub frags: u32,
-    pub deaths: Option<u32>,
-    pub health: Option<u32>,
-    pub secret: bool,
+    pub team: u8,
+    pub deaths: u32,
+    pub skill: u32,
+}
+
+/// A team's details
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct Team {
+    pub name: String,
+    pub score: i32,
 }
 
 /// A query response.
@@ -26,16 +29,14 @@ pub struct Player {
 pub struct Response {
     pub name: String,
     pub map: String,
-    pub map_title: Option<String>,
-    pub admin_contact: Option<String>,
-    pub admin_name: Option<String>,
     pub has_password: bool,
     pub game_type: String,
     pub game_version: String,
     pub players_maximum: usize,
     pub players_online: usize,
-    pub players_minimum: u8,
+    pub players_minimum: Option<u8>,
     pub players: Vec<Player>,
+    pub teams: Vec<Team>,
     pub tournament: bool,
     pub unused_entries: HashMap<String, String>,
 }